MusicMood/requirements.txt
Django>=5.0
psycopg[binary]>=3.1
django-environ>=0.11
djangorestframework>=3.15
django-filter>=24.2
drf-spectacular>=0.27
celery>=5.4
redis>=5.0
httpx>=0.27
pydantic>=2.8
numpy>=1.26
scikit-learn>=1.5
sentence-transformers>=3.0
python-dotenv>=1.0
